各項目のグラフページを作成

+++ db/migrate/20201214030800_create_conditions.rb
@@ -2,8 +2,8 @@ class CreateConditions < ActiveRecord::Migration[6.0]
   def change
     create_table :conditions do |t|
       t.datetime   :start_time, null: false
-      t.integer    :temperature
-      t.integer    :weight
+      t.float      :temperature
+      t.float      :weight
       t.integer    :blood_max
       t.integer    :blood_min
       t.integer    :condition,  null: false
